fix: suppress warnings and handle edge cases
- Use 'dtype' instead of deprecated 'torch_dtype' in model loading
- Suppress CodeCarbon FutureWarning about DataFrame concatenation
- Handle nvidia-smi returning [N/A] for GPU utilization

+++ src/llm_energy_measure/core/model_loader.py
@@ -140,7 +140,7 @@ def load_model_tokenizer(
         else:
             model = AutoModelForCausalLM.from_pretrained(
                 model_name,
-                torch_dtype=dtype,
+                dtype=dtype,  # Use 'dtype' instead of deprecated 'torch_dtype'
                 device_map="auto",
             )
     except Exception as e:
